Add some methods to AIsolatedGameTest to manipulate collections
